scripts/the-internet-herokuapp_v2/requirements.txt
selenium
selenium-wire
pyyaml
pytest
webdriver-manager
os